requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.9
aiohttp>=3.8.0
